samples: bluetooth: radio_coex_1wire: Add support for nRF54H20
This commit adds support for nRF54H20 for this sample.

Explicit binding and extra node is needed because nrf-regtool doesn't
pick up properties from the zephyr,user node. Because of this, it
doesn't add to the UICR of the radio core the gpio that sample needs.

Other platforms don't have such security mechanism and thus don't need
this workaround.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>